+++ site/assets/js/snippets.js
@@ -1,193 +1,15 @@
-// NOTICE!!! Initially embedded in our docs this JavaScript
-// file contains elements that can help you create reproducible
-// use cases in StackBlitz for instance.
-// In a real project please adapt this content to your needs.
-// ++++++++++++++++++++++++++++++++++++++++++
-
-/*!
+/*
  * JavaScript for Bootstrap's docs (https://getbootstrap.com/)
  * Copyright 2011-2024 The Bootstrap Authors
  * Licensed under the Creative Commons Attribution 3.0 Unported License.
  * For details, see https://creativecommons.org/licenses/by/3.0/.
  */
 
-/* global boosted: false */
+// Note that this file is not published; we only include it in scrpts.html
+// for StackBlitz to work
 
-(() => {
-  'use strict'
+/* eslint-disable import/no-unresolved */
+import snippets from 'js/partials/snippets.js'
+/* eslint-enable import/no-unresolved */
 
-  // Boosted mod: Try not using `const` at the first layer in this file, otherwise Storybook will fail to render some examples.
-
-  // storybook-start overview
-  // --------
-  // Tooltips
-  // --------
-  // Instantiate all tooltips in a docs or StackBlitz
-  // storybook-start tooltips
-  document.querySelectorAll('[data-bs-toggle="tooltip"]')
-    .forEach(tooltip => {
-      new boosted.Tooltip(tooltip)
-    })
-  // storybook-end tooltips
-
-  // --------
-  // Popovers
-  // --------
-  // Instantiate all popovers in docs or StackBlitz
-  // storybook-start popovers
-  document.querySelectorAll('[data-bs-toggle="popover"]')
-    .forEach(popover => {
-      new boosted.Popover(popover)
-    })
-  // storybook-end popovers
-  // storybook-end overview
-
-  // -------------------------------
-  // Toasts
-  // -------------------------------
-  // Used by 'Placement' example in docs or StackBlitz
-  // storybook-start toasts
-  if (document.getElementById('toastPlacement')) {
-    document.getElementById('selectToastPlacement').addEventListener('change', function () {
-      if (!document.getElementById('toastPlacement').dataset.originalClass) {
-        document.getElementById('toastPlacement').dataset.originalClass = document.getElementById('toastPlacement').className
-      }
-
-      document.getElementById('toastPlacement').className = `${document.getElementById('toastPlacement').dataset.originalClass} ${this.value}`
-    })
-  }
-
-  // Instantiate all toasts in docs pages only
-  document.querySelectorAll('.bd-example .toast')
-    .forEach(toastNode => {
-      const toast = new boosted.Toast(toastNode, {
-        autohide: false
-      })
-
-      toast.show()
-    })
-
-  // Instantiate all toasts in docs pages only
-  // js-docs-start live-toast
-  if (document.getElementById('liveToastBtn')) {
-    document.getElementById('liveToastBtn').addEventListener('click', () => {
-      boosted.Toast.getOrCreateInstance(document.getElementById('liveToast')).show()
-    })
-  }
-  // js-docs-end live-toast
-  // storybook-end toasts
-
-  // -------------------------------
-  // Alerts
-  // -------------------------------
-  // Used in 'Show live alert' example in docs or StackBlitz
-
-  // Boosted mod: adapted innerHTML to have the icon and so added a parameter within `appendAlert()` and a tooltip on `.btn-close`
-  // storybook-start alerts
-  // js-docs-start live-alert
-  function appendAlert(message, type, typeVisuallyHidden) {
-    const wrapper = document.createElement('div')
-    wrapper.innerHTML = [
-      `<div class="alert alert-${type} alert-dismissible" role="alert">`,
-      '  <span class="alert-icon">',
-      `    <span class="visually-hidden">${typeVisuallyHidden}</span>`,
-      '  </span>',
-      `  <p>${message}</p>`,
-      '  <button type="button" class="btn-close" data-bs-dismiss="alert" data-bs-toggle="tooltip" data-bs-placement="bottom" data-bs-title="Close">',
-      '    <span class="visually-hidden">Close</span>',
-      '  </button>',
-      '</div>'
-    ].join('')
-
-    document.getElementById('liveAlertPlaceholder').append(wrapper)
-
-    // Create `.btn-close` tooltip after `innerHTML` has been modified
-    const btnClose = wrapper.querySelector('.btn-close')
-    const tooltip = new boosted.Tooltip(btnClose)
-    btnClose.addEventListener('click', () => {
-      tooltip.hide()
-    })
-  }
-
-  if (document.getElementById('liveAlertBtn')) {
-    document.getElementById('liveAlertBtn').addEventListener('click', () => {
-      appendAlert('Nice, you triggered this alert message!', 'success', 'Success')
-    })
-  }
-  // js-docs-end live-alert
-  // storybook-end alerts
-
-  // --------
-  // Carousels
-  // --------
-  // Instantiate all non-autoplaying carousels in docs or StackBlitz
-  // storybook-start carousel
-  document.querySelectorAll('.carousel:not([data-bs-ride="carousel"])')
-    .forEach(carousel => {
-      boosted.Carousel.getOrCreateInstance(carousel)
-    })
-  // storybook-end carousel
-
-  // -------------------------------
-  // Checks & Radios
-  // -------------------------------
-  // Indeterminate checkbox example in docs and StackBlitz
-  // storybook-start checksradios
-  document.querySelectorAll('.bd-example-indeterminate [type="checkbox"]')
-    .forEach(checkbox => {
-      if (checkbox.id.includes('Indeterminate')) {
-        checkbox.indeterminate = true
-      }
-    })
-  // storybook-end checksradios
-
-  // -------------------------------
-  // Links
-  // -------------------------------
-  // Disable empty links in docs examples only
-  document.querySelectorAll('.bd-content [href="#"]')
-    .forEach(link => {
-      link.addEventListener('click', event => {
-        event.preventDefault()
-      })
-    })
-
-  // -------------------------------
-  // Modal
-  // -------------------------------
-  // Modal 'Varying modal content' example in docs and StackBlitz
-  // storybook-start modal
-  // js-docs-start varying-modal-content
-  if (document.getElementById('exampleModal')) {
-    document.getElementById('exampleModal').addEventListener('show.bs.modal', event => {
-      // Button that triggered the modal
-      const button = event.relatedTarget
-      // Extract info from data-bs-* attributes
-      const recipient = button.getAttribute('data-bs-whatever')
-      // If necessary, you could initiate an Ajax request here
-      // and then do the updating in a callback.
-
-      // Update the modal's content.
-      const modalTitle = document.getElementById('exampleModal').querySelector('.modal-title')
-      const modalBodyInput = document.getElementById('exampleModal').querySelector('.modal-body input')
-
-      modalTitle.textContent = `New message to ${recipient}`
-      modalBodyInput.value = recipient
-    })
-  }
-  // js-docs-end varying-modal-content
-  // storybook-end modal
-
-  // -------------------------------
-  // Offcanvas
-  // -------------------------------
-  // 'Offcanvas components' example in docs only
-  // storybook-start offcanvas
-  document.querySelectorAll('.bd-example-offcanvas .offcanvas')
-    .forEach(offcanvas => {
-      offcanvas.addEventListener('show.bs.offcanvas', event => {
-        event.preventDefault()
-      }, false)
-    })
-  // storybook-end offcanvas
-})()
+snippets()
