+++ site/assets/scss/_search.scss
@@ -1,5 +1,7 @@
 // stylelint-disable selector-class-pattern
 
+@import "@docsearch/css/dist/style";
+
 // Boosted mod: the entire file is modified and is specific to Boosted
 
 :root {
